Move EncodedStream to messaging
This class needs to stay Java 6 compatible.
