+++ subprojects/core/src/main/java/org/gradle/process/internal/worker/child/BootstrapSecurityManager.java
@@ -17,7 +17,7 @@
 package org.gradle.process.internal.worker.child;
 
 import org.gradle.internal.reflect.JavaReflectionUtil;
-import org.gradle.process.internal.streams.EncodedStream;
+import org.gradle.internal.stream.EncodedStream;
 
 import java.io.DataInputStream;
 import java.io.File;
