+++ subprojects/core/src/main/java/org/gradle/process/internal/worker/child/WorkerImplementationFactory.java
@@ -17,8 +17,8 @@
 package org.gradle.process.internal.worker.child;
 
 import org.gradle.internal.remote.Address;
-import org.gradle.process.internal.worker.DefaultWorkerProcessBuilder;
 import org.gradle.process.internal.JavaExecHandleBuilder;
+import org.gradle.process.internal.worker.WorkerProcessBuilder;
 
 import java.net.URL;
 import java.util.List;
@@ -27,5 +27,5 @@ public interface WorkerImplementationFactory {
     /**
      * Configures the Java command that will be used to launch the child process.
      */
-    void prepareJavaCommand(Object workerId, String displayName, DefaultWorkerProcessBuilder processBuilder, List<URL> implementationClassPath, Address serverAddress, JavaExecHandleBuilder execSpec, boolean publishJvmMemoryInfo);
+    void prepareJavaCommand(Object workerId, String displayName, WorkerProcessBuilder processBuilder, List<URL> implementationClassPath, Address serverAddress, JavaExecHandleBuilder execSpec, boolean publishJvmMemoryInfo);
 }
