+++ subprojects/messaging/src/test/groovy/org/gradle/internal/stream/EncodedStreamTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.process.internal.streams
+package org.gradle.internal.stream
 
 import spock.lang.Specification
